update sprint 2 features

+++ client/src/Views/abonnée-pages/component/EditActPopUp.js
@@ -1,24 +1,27 @@
-import React, { useState, useEffect, useContext } from "react";
+import React, { useState, useContext } from "react";
 import Modal from "react-bootstrap/Modal";
-import Select from "react-select";
-import makeAnimated from "react-select/animated";
-import AddIcon from "@material-ui/icons/Add";
-import RemoveIcon from "@material-ui/icons/Remove";
 import MapIcon from "@material-ui/icons/Map";
 import PubContext from "../../../Context/Publication/pubContext";
+import swal from "sweetalert";
+import moment from "moment";
+import addPubsFormValidation from "../../../utilis/addPubsFromValidation";
 
 function EditActPopUp(props) {
   //app level state
   const pubContext = useContext(PubContext);
   const { editPub, deletePub } = pubContext;
+  //component level state
   const [evenement, setEvenement] = useState({
-    description: "",
+    description: props.actDetail.description,
     adresse: "",
-    date_DebutPub: "",
-    heure_debutPub: "",
-    date_FinPub: "",
-    heure_finPub: "",
+    date_DebutPub: moment(props.actDetail.date_DebutPub).format("YYYY-MM-DD"),
+    heure_debutPub: props.actDetail.heure_debutPub,
+    date_FinPub: moment(props.actDetail.date_FinPub).format("YYYY-MM-DD"),
+    heure_finPub: props.actDetail.heure_finPub,
   });
+  //state for handling error msg
+  const [errorsMsg, setErrorsMsg] = useState({});
+
   const {
     description,
     adresse,
@@ -33,23 +36,52 @@ function EditActPopUp(props) {
       [event.target.name]: event.target.value,
     });
   };
-
   const editActivité = () => {
-    editPub(evenement, props.actDetail._id);
-    setEvenement({
-      description: "",
-      adresse: "",
-      nbr_place: "",
-      date_DebutPub: "",
-      heure_debutPub: "",
-      date_FinPub: "",
-      heure_finPub: "",
-      tarifEvent: "",
-    });
+    setErrorsMsg(addPubsFormValidation(evenement));
+    if (Object.keys(errorsMsg).length <= 0) {
+      editPub(
+        {
+          description,
+          adresse,
+          date_DebutPub: moment(date_DebutPub + " " + heure_debutPub).format(),
+          heure_debutPub,
+          date_FinPub: moment(date_FinPub + " " + heure_finPub).format(),
+          heure_finPub,
+        },
+        props.actDetail._id
+      ).then(
+        setEvenement({
+          description: "",
+          adresse: "",
+          nbr_place: "",
+          date_DebutPub: "",
+          heure_debutPub: "",
+          date_FinPub: "",
+          heure_finPub: "",
+          tarifEvent: "",
+        }),
+        props.onHide()
+      );
+    }
   };
 
   const deleteActivité = () => {
-    deletePub(props.actDetail._id);
+    swal({
+      title: `Vous êtes sûre de supprimer?`,
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        swal(
+          deletePub(props.actDetail._id),
+          swal({ icon: "success", title: "Activité supprimé avec succés!" }),
+          props.onHide()
+        );
+      } else {
+        swal("opération annuler!");
+      }
+    });
   };
   return (
     <Modal
@@ -74,7 +106,7 @@ function EditActPopUp(props) {
                 cols="50"
                 placeholder="Description"
                 name="description"
-                defaultValue={props.actDetail.description}
+                defaultValue={evenement.description}
                 onChange={handelChange}
                 required
               />
@@ -83,18 +115,14 @@ function EditActPopUp(props) {
                 <input
                   type="date"
                   name="date_DebutPub"
-                  defaultValue={
-                    props.actDetail.date_DebutPub
-                      ? props.actDetail.date_DebutPub.substr(0, 10)
-                      : ""
-                  }
+                  defaultValue={evenement.date_DebutPub}
                   onChange={handelChange}
                   required
                 />
                 <input
                   type="time"
                   name="heure_debutPub"
-                  defaultValue={props.actDetail.heure_debutPub}
+                  defaultValue={evenement.heure_debutPub}
                   onChange={handelChange}
                 />
               </div>
@@ -104,18 +132,14 @@ function EditActPopUp(props) {
                 <input
                   type="date"
                   name="date_FinPub"
-                  defaultValue={
-                    props.actDetail.date_FinPub
-                      ? props.actDetail.date_FinPub.substr(0, 10)
-                      : ""
-                  }
+                  defaultValue={evenement.date_FinPub}
                   onChange={handelChange}
                   required
                 />
                 <input
                   type="time"
                   name="heure_finPub"
-                  defaultValue={props.actDetail.heure_finPub}
+                  defaultValue={evenement.heure_finPub}
                   onChange={handelChange}
                 />
               </div>
